poster/requirements.txt
fastapi
uvicorn
requests
langchain
langchain-community
langchain_google_genai 
google-generativeai 
transformers 
diffusers 
numpy 
opencv-python 
Pillow 
torch